# backend/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.30.0

package db

import (
	"database/sql"
	"time"
)

type Download struct {
	ID          string
	Channel     string
	Url         string
	Title       sql.NullString
	Status      string
	Progress    int64
	Filename    sql.NullString
	Error       sql.NullString
	CreatedAt   time.Time
	CompletedAt sql.NullTime
}