# spec/tsp-output/@typespec/openapi3/openapi.yaml
openapi: 3.0.0
info:
  title: Media Downloader API
  version: 0.0.0
tags: []
paths:
  /api/{secret}:
    get:
      operationId: getChannelInfo
      description: チャンネル情報を取得
      parameters:
        - name: secret
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChannelInfoResponse'
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/{secret}/downloads:
    get:
      operationId: listDownloads
      description: ダウンロード一覧を取得
      parameters:
        - name: secret
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DownloadListResponse'
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      operationId: createDownload
      description: ダウンロードを開始
      parameters:
        - name: secret
          in: path
          required: true
          schema:
            type: string
      responses:
        '202':
          description: The request has been accepted for processing, but processing has not yet completed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Download'
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDownloadRequest'
components:
  schemas:
    ChannelInfoResponse:
      type: object
      required:
        - name
      properties:
        name:
          type: string
      description: チャンネル情報レスポンス
    CreateDownloadRequest:
      type: object
      required:
        - url
      properties:
        url:
          type: string
      description: ダウンロード開始リクエスト
    Download:
      type: object
      required:
        - id
        - url
        - status
        - progress
        - createdAt
      properties:
        id:
          type: string
        url:
          type: string
        title:
          type: string
        status:
          $ref: '#/components/schemas/DownloadStatus'
        progress:
          type: integer
          format: int32
        filename:
          type: string
        error:
          type: string
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
      description: ダウンロード情報
    DownloadListResponse:
      type: object
      required:
        - downloads
      properties:
        downloads:
          type: array
          items:
            $ref: '#/components/schemas/Download'
      description: ダウンロード一覧レスポンス
    DownloadStatus:
      type: string
      enum:
        - pending
        - downloading
        - converting
        - completed
        - error
      description: ダウンロードステータス
    ErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
      description: エラーレスポンス